refactor: tokenAmountInputs to pool Form component
